+++ TodoAdd.js
@@ -1,15 +1,7 @@
-import React,{useCallback} from "react";
+import React from "react";
+import useAddTodo from "./UseAddTodo";
 function TodoAdd({todoList,setTodoList}){
-   const addTodo=useCallback((event)=>{
-        event.preventDefault();
-        const newTodoItem={
-            id: Date.now(),
-            text: event.target["todo-text"].value,
-            done:false,
-        };
-        setTodoList([...todoList,newTodoItem]);
-        event.target.reset();
-    },[todoList,setTodoList]);
+    const addTodo=useAddTodo(todoList,setTodoList);
     return(
         <form className="todo-add" onSubmit={addTodo}>
             <input id="input1" type="text" name="todo-text" required placeholder="오늘의 할 일은?"></input>
@@ -18,5 +10,4 @@ function TodoAdd({todoList,setTodoList}){
 
     )
 }
-
 export default React.memo(TodoAdd);
